human-ad-sales-hub/scripts/requirements.txt
googlemaps>=4.10.0
supabase>=2.0.0
playwright>=1.40.0
python-dotenv>=1.0.0
